requirements.txt
pandas>=2.1
openpyxl>=3.1
requests>=2.31
urllib3>=2.0
beautifulsoup4>=4.12
lxml>=4.9
tldextract>=5.1
playwright>=1.41
trafilatura>=1.8
readability-lxml>=0.8

# GitHub Actions automation (Google Drive input + reliability helpers)
google-api-python-client>=2.120
google-auth>=2.28
google-auth-httplib2>=0.2
tenacity>=8.2
gdown>=5.2
